pyproject.toml
[tool.poetry]
name = "personal-crm-app"
version = "0.1.0"
description = "FriendCRM app written in Pynecone"
authors = ["Chris Watkins <chris.watkins93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pynecone = "^0.1.20"
pandas = "^1.5.3"
passlib = "^1.7.4"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = { extras = ["d"], version = "^23.1.0" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
